requirements.txt
requests==2.31.0
fastapi
uvicorn
python-telegram-bot
requests
